refactor: move elderly school models to models/elderly-school/
ตามแบบ models/smart-papar/ — ชื่อ model/collection ไม่เปลี่ยน ข้อมูลเดิมไม่กระทบ
