+++ pages/api/smart-health/elderly/assessments.js
@@ -1,5 +1,5 @@
 import dbConnect from "@/lib/dbConnect";
-import ElderlyMentalHealthAssessment from "@/models/ElderlyMentalHealthAssessment";
+import ElderlyMentalHealthAssessment from "@/models/elderly-school/ElderlyMentalHealthAssessment";
 import { is2QPositive, score9Q } from "@/lib/elderly-school/mentalHealth";
 import { ObjectId } from "mongodb";
